Update test, add icon, create some home page ui
Add mock test for weather api
Add icons for weather conditions
Add some home page ui

+++ lib/src/core/domain/entites/hour.dart
@@ -1,6 +1,6 @@
 class HourEntity {
   final String time;
-  final double temp;
+  final int temp;
   final String conditionImageUrl;
 
   HourEntity({
